fix: provide files to github action; mv server test to server
